refactor(rime): remove unused event type and consolidate tests
